+++ usr.sbin/snmpctl/snmpctl.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: snmpctl.c,v 1.14 2012/09/17 16:43:59 reyk Exp $	*/
+/*	$OpenBSD: snmpctl.c,v 1.15 2013/05/07 09:32:58 jsg Exp $	*/
 
 /*
  * Copyright (c) 2007, 2008 Reyk Floeter <reyk@openbsd.org>
@@ -250,7 +250,7 @@ monitor(struct imsg *imsg)
 	imn = monitor_lookup(imsg->hdr.type);
 	printf("%s: imsg type %u len %u peerid %u pid %d\n", imn->name,
 	    imsg->hdr.type, imsg->hdr.len, imsg->hdr.peerid, imsg->hdr.pid);
-	printf("\ttimestamp: %u, %s", now, ctime(&now));
+	printf("\ttimestamp: %lld, %s", (long long)now, ctime(&now));
 	if (imn->type == -1)
 		done = 1;
 	if (imn->func != NULL)
